feat: 게시물 조회시 댓글 목록 보임(answer부분 다소 변경)

+++ server/src/main/java/fuckingrullet/server/question/repository/QuestionRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
+public interface QuestionRepository extends JpaRepository<Question, Long> {
 
     Optional<Question> findByTitle(String title);
